requirements.txt
selenium
pyspellchecker
webdriver-manager
